+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved `Readable` to `genbase`
 - Use `genbase.SeedMixin` for seeds
 - Use `genbase.internationalization` for internationalization
+- Requires `genbase>=0.1.6`
 
 ## [0.5.5] - 2021-11-17
 ### Changed